// src/utilities/types.ts
export enum Color {
  Black = 'Black',
  Blue = 'Blue',
  Orange = 'Orange',
  White = 'White',
}

export enum Font {
  Montez = 'Montez',
  Regular = 'Regular',
}

export enum TransitionDuration {
  Immediate = 100,
  Quickest = 400,
  Quicker = 600,
  Quick = 800,
  Slow = 1000,
  Slower = 1200,
  Slowest = 1400,
}

export type SlideDirection = 'up' | 'down' | 'left' | 'right';